# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Trading Bot Web3 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TradeSide = (typeof TradeSide)[keyof typeof TradeSide];

export const TradeSide = {
  buy: "buy",
  sell: "sell",
} as const;

export type TradeStatus = (typeof TradeStatus)[keyof typeof TradeStatus];

export const TradeStatus = {
  pending: "pending",
  open: "open",
  closed: "closed",
  settled: "settled",
} as const;

export interface Trade {
  id: number;
  symbol: string;
  side: TradeSide;
  price: string;
  quantity: string;
  /** @nullable */
  pnl?: string | null;
  strategy: string;
  /** @nullable */
  confidence?: number | null;
  status: TradeStatus;
  /** @nullable */
  chainTxHash?: string | null;
  /** @nullable */
  chainSettledAt?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateTradeBodySide =
  (typeof CreateTradeBodySide)[keyof typeof CreateTradeBodySide];

export const CreateTradeBodySide = {
  buy: "buy",
  sell: "sell",
} as const;

export interface CreateTradeBody {
  symbol: string;
  side: CreateTradeBodySide;
  price: string;
  quantity: string;
  strategy: string;
  /** @nullable */
  confidence?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface SettleTradeBody {
  pnl: string;
  chainTxHash?: string;
  walletAddress: string;
}

export interface PortfolioSummary {
  totalValue: string;
  totalPnl: string;
  winRate: number;
  totalTrades: number;
  openTrades: number;
  settledTrades: number;
  onChainSettled: number;
  /** @nullable */
  topSymbol?: string | null;
}

export interface PnlDataPoint {
  date: string;
  pnl: number;
  cumulative: number;
}

export type AgentDecisionAction =
  (typeof AgentDecisionAction)[keyof typeof AgentDecisionAction];

export const AgentDecisionAction = {
  buy: "buy",
  sell: "sell",
  watch: "watch",
  hold: "hold",
} as const;

export interface AgentDecision {
  id: number;
  symbol: string;
  action: AgentDecisionAction;
  confidence: number;
  price: string;
  reasoning: string;
  indicators: string;
  strategy: string;
  createdAt: string;
}

export type CreateAgentDecisionBodyAction =
  (typeof CreateAgentDecisionBodyAction)[keyof typeof CreateAgentDecisionBodyAction];

export const CreateAgentDecisionBodyAction = {
  buy: "buy",
  sell: "sell",
  watch: "watch",
  hold: "hold",
} as const;

export interface CreateAgentDecisionBody {
  symbol: string;
  action: CreateAgentDecisionBodyAction;
  confidence: number;
  price: string;
  reasoning: string;
  indicators: string;
  strategy: string;
}

export type AgentStatusStatus =
  (typeof AgentStatusStatus)[keyof typeof AgentStatusStatus];

export const AgentStatusStatus = {
  running: "running",
  stopped: "stopped",
  paused: "paused",
} as const;

export interface AgentStatus {
  status: AgentStatusStatus;
  strategy: string;
  uptime: number;
  krakenConnected: boolean;
  web3Connected: boolean;
  /** @nullable */
  contractAddress?: string | null;
  network: string;
}

export interface ChainStatus {
  connected: boolean;
  network: string;
  /** @nullable */
  contractAddress?: string | null;
  contractDeployed: boolean;
  rpcUrl: string;
  totalOnChainTrades: number;
  totalOnChainPnl: string;
}

export interface OnChainPnl {
  totalPnl: string;
  totalTrades: number;
  /** @nullable */
  lastUpdated?: string | null;
}

export interface OnChainTrade {
  tradeId: number;
  symbol: string;
  side: string;
  price: string;
  pnl: string;
  txHash: string;
  timestamp: string;
}

export type ListTradesParams = {
  limit?: number;
  /**
   * @nullable
   */
  symbol?: string | null;
};

export type ListAgentDecisionsParams = {
  limit?: number;
};

export type ListOnChainTradesParams = {
  limit?: number;
};