# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Trading Bot Web3 API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all trades
 */
export const listTradesQueryLimitDefault = 50;

export const ListTradesQueryParams = zod.object({
  limit: zod.coerce.number().default(listTradesQueryLimitDefault),
  symbol: zod.coerce.string().nullish(),
});

export const ListTradesResponseItem = zod.object({
  id: zod.number(),
  symbol: zod.string(),
  side: zod.enum(["buy", "sell"]),
  price: zod.string(),
  quantity: zod.string(),
  pnl: zod.string().nullish(),
  strategy: zod.string(),
  confidence: zod.number().nullish(),
  status: zod.enum(["pending", "open", "closed", "settled"]),
  chainTxHash: zod.string().nullish(),
  chainSettledAt: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListTradesResponse = zod.array(ListTradesResponseItem);

/**
 * @summary Record a new trade
 */
export const CreateTradeBody = zod.object({
  symbol: zod.string(),
  side: zod.enum(["buy", "sell"]),
  price: zod.string(),
  quantity: zod.string(),
  strategy: zod.string(),
  confidence: zod.number().nullish(),
  notes: zod.string().nullish(),
});

export const GetTradeParams = zod.object({
  id: zod.coerce.number(),
});

export const GetTradeResponse = zod.object({
  id: zod.number(),
  symbol: zod.string(),
  side: zod.enum(["buy", "sell"]),
  price: zod.string(),
  quantity: zod.string(),
  pnl: zod.string().nullish(),
  strategy: zod.string(),
  confidence: zod.number().nullish(),
  status: zod.enum(["pending", "open", "closed", "settled"]),
  chainTxHash: zod.string().nullish(),
  chainSettledAt: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Settle trade on-chain via smart contract
 */
export const SettleTradeParams = zod.object({
  id: zod.coerce.number(),
});

export const SettleTradeBody = zod.object({
  pnl: zod.string(),
  chainTxHash: zod.string().optional(),
  walletAddress: zod.string(),
});

export const SettleTradeResponse = zod.object({
  id: zod.number(),
  symbol: zod.string(),
  side: zod.enum(["buy", "sell"]),
  price: zod.string(),
  quantity: zod.string(),
  pnl: zod.string().nullish(),
  strategy: zod.string(),
  confidence: zod.number().nullish(),
  status: zod.enum(["pending", "open", "closed", "settled"]),
  chainTxHash: zod.string().nullish(),
  chainSettledAt: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Get portfolio summary stats
 */
export const GetPortfolioSummaryResponse = zod.object({
  totalValue: zod.string(),
  totalPnl: zod.string(),
  winRate: zod.number(),
  totalTrades: zod.number(),
  openTrades: zod.number(),
  settledTrades: zod.number(),
  onChainSettled: zod.number(),
  topSymbol: zod.string().nullish(),
});

/**
 * @summary Get P&L history over time
 */
export const GetPnlHistoryResponseItem = zod.object({
  date: zod.string(),
  pnl: zod.number(),
  cumulative: zod.number(),
});
export const GetPnlHistoryResponse = zod.array(GetPnlHistoryResponseItem);

/**
 * @summary List recent AI agent decisions
 */
export const listAgentDecisionsQueryLimitDefault = 20;

export const ListAgentDecisionsQueryParams = zod.object({
  limit: zod.coerce.number().default(listAgentDecisionsQueryLimitDefault),
});

export const ListAgentDecisionsResponseItem = zod.object({
  id: zod.number(),
  symbol: zod.string(),
  action: zod.enum(["buy", "sell", "watch", "hold"]),
  confidence: zod.number(),
  price: zod.string(),
  reasoning: zod.string(),
  indicators: zod.string(),
  strategy: zod.string(),
  createdAt: zod.string(),
});
export const ListAgentDecisionsResponse = zod.array(
  ListAgentDecisionsResponseItem,
);

/**
 * @summary Record an AI agent decision
 */
export const CreateAgentDecisionBody = zod.object({
  symbol: zod.string(),
  action: zod.enum(["buy", "sell", "watch", "hold"]),
  confidence: zod.number(),
  price: zod.string(),
  reasoning: zod.string(),
  indicators: zod.string(),
  strategy: zod.string(),
});

/**
 * @summary Get current agent operational status
 */
export const GetAgentStatusResponse = zod.object({
  status: zod.enum(["running", "stopped", "paused"]),
  strategy: zod.string(),
  uptime: zod.number(),
  krakenConnected: zod.boolean(),
  web3Connected: zod.boolean(),
  contractAddress: zod.string().nullish(),
  network: zod.string(),
});

/**
 * @summary Get Web3/blockchain connection status and contract info
 */
export const GetChainStatusResponse = zod.object({
  connected: zod.boolean(),
  network: zod.string(),
  contractAddress: zod.string().nullish(),
  contractDeployed: zod.boolean(),
  rpcUrl: zod.string(),
  totalOnChainTrades: zod.number(),
  totalOnChainPnl: zod.string(),
});

/**
 * @summary Fetch total P&L recorded on the smart contract
 */
export const GetOnChainPnlResponse = zod.object({
  totalPnl: zod.string(),
  totalTrades: zod.number(),
  lastUpdated: zod.string().nullish(),
});

/**
 * @summary List trades recorded on the smart contract
 */
export const listOnChainTradesQueryLimitDefault = 20;

export const ListOnChainTradesQueryParams = zod.object({
  limit: zod.coerce.number().default(listOnChainTradesQueryLimitDefault),
});

export const ListOnChainTradesResponseItem = zod.object({
  tradeId: zod.number(),
  symbol: zod.string(),
  side: zod.string(),
  price: zod.string(),
  pnl: zod.string(),
  txHash: zod.string(),
  timestamp: zod.string(),
});
export const ListOnChainTradesResponse = zod.array(
  ListOnChainTradesResponseItem,
);