# lib/api-zod/src/generated/types/agentDecision.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Trading Bot Web3 API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentDecisionAction } from "./agentDecisionAction";

export interface AgentDecision {
  id: number;
  symbol: string;
  action: AgentDecisionAction;
  confidence: number;
  price: string;
  reasoning: string;
  indicators: string;
  strategy: string;
  createdAt: string;
}