# lib/api-zod/src/generated/types/agentStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Trading Bot Web3 API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentStatusStatus } from "./agentStatusStatus";

export interface AgentStatus {
  status: AgentStatusStatus;
  strategy: string;
  uptime: number;
  krakenConnected: boolean;
  web3Connected: boolean;
  /** @nullable */
  contractAddress?: string | null;
  network: string;
}