# lib/api-zod/src/generated/types/createAgentDecisionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Trading Bot Web3 API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAgentDecisionBodyAction } from "./createAgentDecisionBodyAction";

export interface CreateAgentDecisionBody {
  symbol: string;
  action: CreateAgentDecisionBodyAction;
  confidence: number;
  price: string;
  reasoning: string;
  indicators: string;
  strategy: string;
}