# lib/api-zod/src/generated/types/onChainTrade.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Trading Bot Web3 API
 * OpenAPI spec version: 0.1.0
 */

export interface OnChainTrade {
  tradeId: number;
  symbol: string;
  side: string;
  price: string;
  pnl: string;
  txHash: string;
  timestamp: string;
}