# lib/api-zod/src/generated/types/pnlDataPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Trading Bot Web3 API
 * OpenAPI spec version: 0.1.0
 */

export interface PnlDataPoint {
  date: string;
  pnl: number;
  cumulative: number;
}